refactor(php): convert tests to test suites (security, usability)
