--- lib/php/src/tests/Feature/SecurityTest.php
@@ -1,22 +1,0 @@
-<?php
-
-namespace Tests\Feature;
-
-//use Illuminate\Foundation\Testing\RefreshDatabase;
-use Tests\TestCase;
-
-class SecurityTest extends TestCase
-{
-    /**
-     * A basic test to check if the root route returns "Hello SecDim".
-     *
-     * @return void
-     */
-    public function testHelloSecDim()
-    {
-        $response = $this->get('/');
-
-        $response->assertStatus(200);
-        $response->assertSeeText('Hello SecDim');
-    }
-}
